Rewrote function in Db and mockDB
